Rename HasTestDir to HasChainStoreDir
