+++ crates/celestia/celestia-integration-tests/src/contexts/bootstrap.rs
@@ -20,10 +20,10 @@ use hermes_cosmos_test_components::bootstrap::traits::chain::build_chain::CanBui
 use hermes_cosmos_test_components::bootstrap::traits::chain::build_chain::ChainFromBootstrapParamsBuilder;
 use hermes_cosmos_test_components::bootstrap::traits::fields::chain_command_path::ChainCommandPathGetter;
 use hermes_cosmos_test_components::bootstrap::traits::fields::chain_command_path::HasChainCommandPath;
+use hermes_cosmos_test_components::bootstrap::traits::fields::chain_store_dir::HasChainStoreDir;
+use hermes_cosmos_test_components::bootstrap::traits::fields::chain_store_dir::TestDirGetter;
 use hermes_cosmos_test_components::bootstrap::traits::fields::random_id::HasRandomIdFlag;
 use hermes_cosmos_test_components::bootstrap::traits::fields::random_id::RandomIdFlagGetter;
-use hermes_cosmos_test_components::bootstrap::traits::fields::test_dir::HasTestDir;
-use hermes_cosmos_test_components::bootstrap::traits::fields::test_dir::TestDirGetter;
 use hermes_cosmos_test_components::bootstrap::traits::generator::generate_wallet_config::WalletConfigGeneratorComponent;
 use hermes_cosmos_test_components::bootstrap::traits::modifiers::modify_comet_config::CanModifyCometConfig;
 use hermes_cosmos_test_components::bootstrap::traits::modifiers::modify_comet_config::CometConfigModifier;
@@ -94,8 +94,8 @@ impl ProvideRuntime<CelestiaAppBootstrap> for CelestiaAppBootstrapComponents {
 }
 
 impl TestDirGetter<CelestiaAppBootstrap> for CelestiaAppBootstrapComponents {
-    fn test_dir(bootstrap: &CelestiaAppBootstrap) -> &PathBuf {
-        &bootstrap.cosmos_bootstrap.test_dir()
+    fn chain_store_dir(bootstrap: &CelestiaAppBootstrap) -> &PathBuf {
+        &bootstrap.cosmos_bootstrap.chain_store_dir()
     }
 }
