+++ crates/cosmos/cosmos-integration-tests/src/contexts/bootstrap.rs
@@ -22,8 +22,8 @@ use hermes_cosmos_test_components::bootstrap::impls::types::genesis_config::Prov
 use hermes_cosmos_test_components::bootstrap::impls::types::wallet_config::ProvideCosmosWalletConfigType;
 use hermes_cosmos_test_components::bootstrap::traits::chain::build_chain::ChainFromBootstrapParamsBuilder;
 use hermes_cosmos_test_components::bootstrap::traits::fields::chain_command_path::ChainCommandPathGetter;
+use hermes_cosmos_test_components::bootstrap::traits::fields::chain_store_dir::TestDirGetter;
 use hermes_cosmos_test_components::bootstrap::traits::fields::random_id::RandomIdFlagGetter;
-use hermes_cosmos_test_components::bootstrap::traits::fields::test_dir::TestDirGetter;
 use hermes_cosmos_test_components::bootstrap::traits::generator::generate_wallet_config::WalletConfigGeneratorComponent;
 use hermes_cosmos_test_components::bootstrap::traits::modifiers::modify_comet_config::CometConfigModifier;
 use hermes_cosmos_test_components::bootstrap::traits::modifiers::modify_genesis_config::CosmosGenesisConfigModifier;
@@ -62,7 +62,7 @@ pub struct CosmosBootstrap {
     pub runtime: HermesRuntime,
     pub builder: Arc<CosmosBuilder>,
     pub should_randomize_identifiers: bool,
-    pub test_dir: PathBuf,
+    pub chain_store_dir: PathBuf,
     pub chain_command_path: PathBuf,
     pub account_prefix: String,
     pub staking_denom: Denom,
@@ -232,8 +232,8 @@ impl ProvideRuntime<CosmosBootstrap> for CosmosBootstrapComponents {
 }
 
 impl TestDirGetter<CosmosBootstrap> for CosmosBootstrapComponents {
-    fn test_dir(bootstrap: &CosmosBootstrap) -> &PathBuf {
-        &bootstrap.test_dir
+    fn chain_store_dir(bootstrap: &CosmosBootstrap) -> &PathBuf {
+        &bootstrap.chain_store_dir
     }
 }
