+++ crates/cosmos/cosmos-integration-tests/tests/celestia.rs
@@ -32,7 +32,7 @@ fn celestia_integration_tests() -> Result<(), Error> {
         runtime: runtime.clone(),
         builder: builder.clone(),
         should_randomize_identifiers: true,
-        test_dir: "./test-data".into(),
+        chain_store_dir: "./test-data".into(),
         chain_command_path: "celestia-appd".into(),
         account_prefix: "celestia".into(),
         compat_mode: Some(CompatMode::V0_34),
@@ -46,7 +46,7 @@ fn celestia_integration_tests() -> Result<(), Error> {
         runtime,
         builder,
         should_randomize_identifiers: true,
-        test_dir: "./test-data".into(),
+        chain_store_dir: "./test-data".into(),
         chain_command_path: "gaiad".into(),
         account_prefix: "cosmos".into(),
         compat_mode: None,
