+++ crates/cosmos/cosmos-test-components/src/bootstrap/components/cosmos_sdk_legacy.rs
@@ -29,9 +29,9 @@ use crate::bootstrap::impls::genesis_legacy::collect_gentxs::LegacyCollectCosmos
 use crate::bootstrap::traits::chain::build_chain::ChainFromBootstrapParamsBuilder;
 use crate::bootstrap::traits::chain::start_chain::ChainFullNodeStarterComponent;
 use crate::bootstrap::traits::fields::chain_command_path::ChainCommandPathGetter;
+use crate::bootstrap::traits::fields::chain_store_dir::TestDirGetter;
 use crate::bootstrap::traits::fields::hd_path::WalletHdPathComponent;
 use crate::bootstrap::traits::fields::random_id::RandomIdFlagGetter;
-use crate::bootstrap::traits::fields::test_dir::TestDirGetter;
 use crate::bootstrap::traits::generator::generate_chain_id::ChainIdGeneratorComponent;
 use crate::bootstrap::traits::generator::generate_wallet_config::WalletConfigGenerator;
 use crate::bootstrap::traits::genesis::add_genesis_account::GenesisAccountAdderComponent;
