+++ crates/cosmos/cosmos-test-components/src/bootstrap/impls/initializers/create_chain_home_dir.rs
@@ -7,7 +7,7 @@ use hermes_relayer_components::runtime::traits::runtime::HasRuntime;
 use hermes_test_components::chain_driver::traits::types::chain::HasChainType;
 use hermes_test_components::runtime::traits::create_dir::CanCreateDir;
 
-use crate::bootstrap::traits::fields::test_dir::HasTestDir;
+use crate::bootstrap::traits::fields::chain_store_dir::HasChainStoreDir;
 use crate::bootstrap::traits::initializers::init_chain_home_dir::ChainHomeDirInitializer;
 
 pub struct CreateChainHomeDirFromTestDir;
@@ -18,7 +18,7 @@ where
     Bootstrap: HasChainType<Chain = Chain>
         + HasRuntime<Runtime = Runtime>
         + CanRaiseError<Runtime::Error>
-        + HasTestDir,
+        + HasChainStoreDir,
     Runtime: CanCreateDir,
     Chain: HasChainIdType,
     Chain::ChainId: Display,
@@ -27,9 +27,9 @@ where
         bootstrap: &Bootstrap,
         chain_id: &Chain::ChainId,
     ) -> Result<Runtime::FilePath, Bootstrap::Error> {
-        let test_dir = bootstrap.test_dir();
+        let chain_store_dir = bootstrap.chain_store_dir();
         let chain_home_dir = Runtime::join_file_path(
-            test_dir,
+            chain_store_dir,
             &Runtime::file_path_from_string(&chain_id.to_string()),
         );
