+++ crates/cosmos/cosmos-test-components/src/bootstrap/traits/fields/chain_store_dir.rs
@@ -3,9 +3,9 @@ use hermes_relayer_components::runtime::traits::runtime::HasRuntime;
 use hermes_test_components::runtime::traits::types::file_path::{FilePath, HasFilePathType};
 
 #[derive_component(TestDirComponent, TestDirGetter<Bootstrap>)]
-pub trait HasTestDir: HasRuntime
+pub trait HasChainStoreDir: HasRuntime
 where
     Self::Runtime: HasFilePathType,
 {
-    fn test_dir(&self) -> &FilePath<Self::Runtime>;
+    fn chain_store_dir(&self) -> &FilePath<Self::Runtime>;
 }
