+++ crates/cosmos/cosmos-test-components/src/bootstrap/traits/fields/mod.rs
@@ -1,4 +1,4 @@
 pub mod chain_command_path;
+pub mod chain_store_dir;
 pub mod hd_path;
 pub mod random_id;
-pub mod test_dir;
